refactor(achievements): enforce public barrel exports pattern
