+++ src/app/(dashboard)/dashboard/management/manage-achievements/bulk-issue/page.tsx
@@ -1,4 +1,4 @@
-import { BulkIssuePanel } from "@/features/achievements/components";
+import { BulkIssuePanel } from "@/features/achievements";
 import { ROLES } from "@/lib/auth/roles";
 import { requireRole } from "@/lib/auth/server";
