+++ src/app/(dashboard)/dashboard/management/manage-achievements/issue/page.tsx
@@ -1,4 +1,4 @@
-import { IssueRevokePanel } from "@/features/achievements/components";
+import { IssueRevokePanel } from "@/features/achievements";
 import { ROLES } from "@/lib/auth/roles";
 import { requireRole } from "@/lib/auth/server";
