+++ src/app/(dashboard)/dashboard/management/manage-achievements/rules/page.tsx
@@ -1,4 +1,4 @@
-import { RulesTable } from "@/features/achievements/components";
+import { RulesTable } from "@/features/achievements";
 import { ROLES } from "@/lib/auth/roles";
 import { requireRole } from "@/lib/auth/server";
