+++ src/features/achievements/components/index.ts
@@ -8,5 +8,6 @@ export { BulkIssueDialog } from "./bulk-issue-dialog";
 export { BulkIssuePanel } from "./bulk-issue-panel";
 export { IssueRevokePanel } from "./issue-revoke-panel";
 export { IssueVCDialog } from "./issue-vc-dialog";
+export { IssueVCDialogListener } from "./issue-vc-dialog-listener";
 export { RulesTable } from "./rules-table";
 export { SimulationPanel } from "./simulation-panel";
