+++ src/features/achievements/components/rule-form-dialog.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import { zodResolver } from "@hookform/resolvers/zod";
-import { Trash2 } from "lucide-react";
 import * as React from "react";
 import { type Resolver, useForm } from "react-hook-form";
 import { z } from "zod";
@@ -16,14 +15,11 @@ import {
 import {
   Form,
   FormControl,
-  FormDescription,
   FormField,
   FormItem,
   FormLabel,
   FormMessage,
 } from "@/components/ui/form";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
 import {
   Select,
   SelectContent,
@@ -32,17 +28,22 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { useInterestGroupsList } from "@/features/interest-groups";
-import {
-  MILESTONE_TYPE_OPTIONS,
-  RULE_TYPES,
-  STREAK_TYPE_OPTIONS,
-} from "../constants/constants";
+import { RULE_TYPES } from "../constants/constants";
 import {
   useCreateRule,
   useUpdateRule,
 } from "../hooks/use-achievement-mutations";
 import { useAchievements } from "../hooks/use-achievements";
 import type { AchievementRule } from "../schemas";
+import {
+  CustomFields,
+  EventFields,
+  IGKarmaFields,
+  MilestoneFields,
+  SkillFields,
+  StreakFields,
+  TaskCompletionFields,
+} from "./rule-form-fields";
 
 // ==========================================
 // Form Schema — flat fields for all 6 rule types
@@ -509,362 +510,34 @@ export function RuleFormDialog({
 
             {/* ── ig_karma ── */}
             {watchedRuleType === "ig_karma" && (
-              <div className="space-y-4">
-                <FormField
-                  control={form.control}
-                  name="ig_id"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Interest Group</FormLabel>
-                      <Select
-                        value={(field.value ?? "") as string}
-                        onValueChange={field.onChange}
-                      >
-                        <FormControl>
-                          <SelectTrigger className="w-full">
-                            <SelectValue
-                              placeholder={
-                                isLoadingIgs
-                                  ? "Loading interest groups…"
-                                  : "Select interest group"
-                              }
-                            />
-                          </SelectTrigger>
-                        </FormControl>
-                        <SelectContent>
-                          {igs.map((ig) => (
-                            <SelectItem key={ig.id} value={ig.id}>
-                              {ig.name}
-                            </SelectItem>
-                          ))}
-                        </SelectContent>
-                      </Select>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={form.control}
-                  name="ig_required_karma"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Required Karma</FormLabel>
-                      <FormControl>
-                        <Input
-                          type="number"
-                          min={0}
-                          placeholder="e.g. 500"
-                          {...field}
-                          value={field.value ?? ""}
-                        />
-                      </FormControl>
-                      <FormDescription>
-                        User must have ≥ this much karma in the selected IG.
-                      </FormDescription>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </div>
+              <IGKarmaFields igs={igs} isLoadingIgs={isLoadingIgs} />
             )}
 
             {/* ── skill ── */}
-            {watchedRuleType === "skill" && (
-              <div className="space-y-4">
-                <FormField
-                  control={form.control}
-                  name="skill_id"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Skill ID</FormLabel>
-                      <FormControl>
-                        <Input
-                          placeholder="UUID of the skill"
-                          {...field}
-                          value={field.value ?? ""}
-                        />
-                      </FormControl>
-                      <FormDescription>
-                        Paste the UUID of the skill from the skills list.
-                      </FormDescription>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={form.control}
-                  name="skill_required_tasks"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Required Tasks</FormLabel>
-                      <FormControl>
-                        <Input
-                          type="number"
-                          min={1}
-                          placeholder="e.g. 10"
-                          {...field}
-                          value={field.value ?? ""}
-                        />
-                      </FormControl>
-                      <FormDescription>
-                        User must have completed ≥ this many tasks tagged with
-                        the skill.
-                      </FormDescription>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </div>
-            )}
+            {watchedRuleType === "skill" && <SkillFields />}
 
             {/* ── streak ── */}
-            {watchedRuleType === "streak" && (
-              <div className="space-y-4">
-                <FormField
-                  control={form.control}
-                  name="streak_type"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Streak Type</FormLabel>
-                      <Select
-                        value={(field.value ?? "") as string}
-                        onValueChange={field.onChange}
-                      >
-                        <FormControl>
-                          <SelectTrigger className="w-full">
-                            <SelectValue placeholder="Select streak type" />
-                          </SelectTrigger>
-                        </FormControl>
-                        <SelectContent>
-                          {STREAK_TYPE_OPTIONS.map((opt) => (
-                            <SelectItem key={opt.value} value={opt.value}>
-                              {opt.label}
-                            </SelectItem>
-                          ))}
-                        </SelectContent>
-                      </Select>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={form.control}
-                  name="streak_required"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Required Streak (days)</FormLabel>
-                      <FormControl>
-                        <Input
-                          type="number"
-                          min={1}
-                          placeholder="e.g. 30"
-                          {...field}
-                          value={field.value ?? ""}
-                        />
-                      </FormControl>
-                      <FormDescription>
-                        User's current streak must be ≥ this number of
-                        consecutive days.
-                      </FormDescription>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </div>
-            )}
+            {watchedRuleType === "streak" && <StreakFields />}
 
             {/* ── milestone ── */}
-            {watchedRuleType === "milestone" && (
-              <div className="space-y-4">
-                <FormField
-                  control={form.control}
-                  name="milestone_type"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Milestone Type</FormLabel>
-                      <Select
-                        value={(field.value ?? "") as string}
-                        onValueChange={field.onChange}
-                      >
-                        <FormControl>
-                          <SelectTrigger className="w-full">
-                            <SelectValue placeholder="Select milestone type" />
-                          </SelectTrigger>
-                        </FormControl>
-                        <SelectContent>
-                          {MILESTONE_TYPE_OPTIONS.map((opt) => (
-                            <SelectItem key={opt.value} value={opt.value}>
-                              {opt.label}
-                            </SelectItem>
-                          ))}
-                        </SelectContent>
-                      </Select>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={form.control}
-                  name="milestone_required_value"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Required Value</FormLabel>
-                      <FormControl>
-                        <Input
-                          type="number"
-                          min={0}
-                          placeholder="e.g. 1000"
-                          {...field}
-                          value={field.value ?? ""}
-                        />
-                      </FormControl>
-                      <FormDescription>
-                        User's aggregate value must be ≥ this number.
-                      </FormDescription>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </div>
-            )}
+            {watchedRuleType === "milestone" && <MilestoneFields />}
 
             {/* ── event ── */}
-            {watchedRuleType === "event" && (
-              <div className="space-y-4">
-                <FormField
-                  control={form.control}
-                  name="event_name"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Event Name</FormLabel>
-                      <FormControl>
-                        <Input
-                          placeholder='e.g. "MuLearn Summit 2024"'
-                          {...field}
-                          value={field.value ?? ""}
-                        />
-                      </FormControl>
-                      <FormDescription>
-                        Exact name of the event as it appears in karma logs.
-                      </FormDescription>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-                <FormField
-                  control={form.control}
-                  name="event_required_attendance"
-                  render={({ field }) => (
-                    <FormItem>
-                      <FormLabel>Required Attendance</FormLabel>
-                      <FormControl>
-                        <Input
-                          type="number"
-                          min={1}
-                          placeholder="e.g. 1"
-                          {...field}
-                          value={field.value ?? ""}
-                        />
-                      </FormControl>
-                      <FormDescription>
-                        User must have ≥ this many approved karma logs for the
-                        event. Usually 1.
-                      </FormDescription>
-                      <FormMessage />
-                    </FormItem>
-                  )}
-                />
-              </div>
-            )}
+            {watchedRuleType === "event" && <EventFields />}
 
             {/* ── task_completion ── */}
-            {watchedRuleType === "task_completion" && (
-              <FormField
-                control={form.control}
-                name="task_hashtag"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Task Hashtag</FormLabel>
-                    <FormControl>
-                      <Input
-                        placeholder='e.g. "#onboarding"'
-                        {...field}
-                        value={field.value ?? ""}
-                      />
-                    </FormControl>
-                    <FormDescription>
-                      User must have at least one approved karma log for a task
-                      with this hashtag.
-                    </FormDescription>
-                    <FormMessage />
-                  </FormItem>
-                )}
-              />
-            )}
+            {watchedRuleType === "task_completion" && <TaskCompletionFields />}
 
             {/* ── Custom ── */}
             {isCustomType && (
-              <div className="space-y-4">
-                <div className="space-y-2">
-                  <Label>Custom Rule Type Name</Label>
-                  <Input
-                    placeholder="Enter custom rule type (e.g. hackathon)"
-                    value={customTypeValue}
-                    onChange={(e) => setCustomTypeValue(e.target.value)}
-                  />
-                </div>
-
-                <div className="space-y-2">
-                  <Label className="flex items-center justify-between">
-                    <span>Conditions</span>
-                    <button
-                      type="button"
-                      onClick={addPair}
-                      className="text-xs text-brand-blue hover:text-brand-blue/80 font-medium transition-colors"
-                    >
-                      + Add field
-                    </button>
-                  </Label>
-                  {customPairs.length === 0 ? (
-                    <p className="text-xs text-muted-foreground italic border border-dashed rounded-lg p-3 text-center">
-                      No conditions yet. Click "+ Add field" to add one.
-                    </p>
-                  ) : (
-                    <div className="space-y-2">
-                      {customPairs.map((pair) => (
-                        <div key={pair.id} className="flex items-center gap-2">
-                          <Input
-                            placeholder="key"
-                            value={pair.key}
-                            onChange={(e) =>
-                              updatePair(pair.id, "key", e.target.value)
-                            }
-                            className="flex-1 font-mono text-xs h-8"
-                          />
-                          <span className="text-muted-foreground text-xs shrink-0">
-                            :
-                          </span>
-                          <Input
-                            placeholder="value"
-                            value={pair.value}
-                            onChange={(e) =>
-                              updatePair(pair.id, "value", e.target.value)
-                            }
-                            className="flex-1 font-mono text-xs h-8"
-                          />
-                          <button
-                            type="button"
-                            onClick={() => removePair(pair.id)}
-                            className="shrink-0 text-muted-foreground hover:text-destructive transition-colors"
-                          >
-                            <Trash2 className="h-3.5 w-3.5" />
-                          </button>
-                        </div>
-                      ))}
-                    </div>
-                  )}
-                </div>
-              </div>
+              <CustomFields
+                customTypeValue={customTypeValue}
+                setCustomTypeValue={setCustomTypeValue}
+                customPairs={customPairs}
+                addPair={addPair}
+                removePair={removePair}
+                updatePair={updatePair}
+              />
             )}
 
             <DialogFooter>
