+++ src/features/achievements/index.ts
@@ -1,15 +1,24 @@
 export {
   activateRule,
+  bulkClaimAchievements,
   bulkIssueAchievements,
+  claimAchievement,
   createAchievement,
   createRule,
   deactivateRule,
   debugAchievement,
   deleteAchievement,
   downloadBulkTemplate,
+  fetchAchievementProgress,
   fetchAchievements,
+  fetchAllAchievementsForUser,
   fetchAuditLogs,
+  fetchEligibleAchievements,
+  fetchIssuedLogs,
   fetchRules,
+  fetchSingleRule,
+  fetchUserAchievements,
+  issueVC,
   manualIssue,
   revokeAchievement,
   simulateForUser,
@@ -24,28 +33,39 @@ export {
   AchievementsTable,
   AuditLogsTable,
   BulkIssueDialog,
+  BulkIssuePanel,
   IssueRevokePanel,
+  IssueVCDialog,
+  IssueVCDialogListener,
   RulesTable,
   SimulationPanel,
 } from "./components";
 export {
   ACHIEVEMENT_KEYS,
+  useAchievementProgress,
   useAchievementQueryClient,
   useAchievementRules,
   useAchievements,
   useActivateRule,
+  useAllAchievementsForUser,
   useAuditLogs,
+  useBulkClaimAchievements,
   useBulkIssue,
+  useClaimAchievement,
   useCreateAchievement,
   useCreateRule,
   useDeactivateRule,
   useDeleteAchievement,
+  useEligibleAchievements,
+  useIssueVC,
   useManualIssue,
   useRevokeAchievement,
   useRules,
   useSimulation,
+  useSingleRule,
   useUpdateAchievement,
   useUpdateRule,
+  useUserAchievements,
 } from "./hooks";
 export type {
   Achievement,
@@ -54,6 +74,7 @@ export type {
   AuditLog,
   CreateAchievementRequest,
   CreateRuleRequest,
+  EligibleAchievement,
   IssuedLog,
   ManualIssueRequest,
   PaginatedIssuedLog,
@@ -74,6 +95,8 @@ export {
   CreateAchievementRequestSchema,
   CreateRuleRequestSchema,
   DjangoResponseSchema,
+  EligibleAchievementListResponseSchema,
+  EligibleAchievementSchema,
   GenericSuccessResponseSchema,
   IssuedLogListResponseSchema,
   IssuedLogSchema,
